Fix -Wsign-compare warning

+++ lib/modular_arithmetic/montgomery_arithmetic/include/hurchalla/montgomery_arithmetic/detail/experimental/montgomery_pow_kary.h
@@ -157,7 +157,7 @@ typename MF::MontgomeryValue montgomery_pow_kary(const MF& mf, typename MF::Mont
 
     while (shift >= static_cast<int>(P)) {
         if (USE_SLIDING_WINDOW_OPTIMIZATION) {
-            while (shift > P && (static_cast<size_t>(n >> (shift-1)) & 1) == 0) {
+            while (shift > static_cast<int>(P) && (static_cast<size_t>(n >> (shift-1)) & 1) == 0) {
                 result = mf.square(result);
                 --shift;
             }
